auth_project/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
asgiref==3.11.0
    # via django
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
cryptography==46.0.3
    # via jwt
django==6.0.1
    # via
    #   djangorestframework
    #   drf-example
djangorestframework==3.16.1
    # via drf-example
iniconfig==2.3.0
    # via pytest
jwt==1.4.0
    # via drf-example
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via pytest
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via pytest-django
pytest-django==4.11.1
    # via drf-example
python-dotenv==1.2.1
    # via drf-example
pytz==2025.2
    # via drf-example
ruff==0.14.14
sqlparse==0.5.5
    # via django
tzdata==2025.3 ; sys_platform == 'win32'
    # via django
